requirements.txt
PyPDF2 == 3.0.1
python-dotenv == 1.0.0  
reportlab == 4.2.5
Pillow == 11.1.0
torch == 2.5.1
torchvision == 0.20.1
easyocr == 1.7.2
numpy == 1.26.4
PyMuPDF == 1.24.13
openai == 2.2.0
asyncpg == 0.31.0
